// src/app/08_SIR.RRHH.Reportes/interfaces/PadronFuncionario.interface.ts
export interface PadronFuncionario {
    id:                 number;
    nroFuncionario:     string;
    apellidos:          string;
    nombres:            string;
    sector:             string;
    tipoRemuneracion:   string;
    horasTrabajadas:    number;
    diasTrabajados:     string;
    ultimaModificacion: Date;   
}